requirements.txt
# ── Core ──────────────────────────────────────────────────────────────────────
pandas>=2.1.0
openpyxl>=3.1.2

# ── HTTP / Async ──────────────────────────────────────────────────────────────
httpx>=0.27.0

# ── Environment variables ─────────────────────────────────────────────────────
python-dotenv>=1.0.0

# ── Optional: progress bar (comment out if not desired) ───────────────────────
# tqdm>=4.66.0
